add source/header for ChunkResolver
